refactor: web에서 로그인이 필요하다는 에러를 http 상태코드를 표현하면서 로그인페이지로 이동하도록 수정
ControllerAdvice에 제한범위 설정
Api의 Result를 반환하는 Result 객체가 api에 있는게 더 맞을 것 같아 이동
