+++ README.md
@@ -134,6 +134,7 @@
 2. `HttpSessionUtils` 클래스를 생성하였습니다. 원하는 대로 수정했습니다. (`Optional`로 리턴) 로그아웃도 `HttpSessionUtils`에서 처리하도록하였습니다.
 3. 에러처리를 위해 Exception 처리에 관련된 내용을 찾아보고 추가하였습니다. 절반은 이해하고 절반은 이해하지 못한 코드지만 사용해보면서 익혀보려고 합니다.
 4. 로그인 하지 않은 경우 에러처리를 위해 401 에러코드를 추가하였습니다.
+5. Html로 Error 코드를 주는 방법을 고민해서 적용하였습니다.
 
 # STEP5
