+++ src/main/java/com/codessquad/qna/api/question/ApiAnswerController.java
@@ -1,5 +1,6 @@
 package com.codessquad.qna.api.question;
 
+import com.codessquad.qna.api.result.Result;
 import com.codessquad.qna.common.error.exception.LoginRequiredException;
 import com.codessquad.qna.common.utils.HttpSessionUtils;
 import com.codessquad.qna.web.question.*;
@@ -7,7 +8,6 @@ import com.codessquad.qna.web.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import javax.security.auth.login.LoginException;
 import javax.servlet.http.HttpSession;
 
 @RestController
