+++ src/main/java/com/codessquad/qna/common/constants/CommonConstants.java
@@ -3,7 +3,6 @@ package com.codessquad.qna.common.constants;
 public class CommonConstants {
 
     public static final String SESSION_LOGIN_USER = "loginUser";
-    public static final String REDIRECT_LOGIN_PAGE = "redirect:/users/login";
 
     private CommonConstants() {}
 }
