+++ src/main/java/com/codessquad/qna/common/constants/ErrorConstants.java
@@ -3,6 +3,7 @@ package com.codessquad.qna.common.constants;
 public class ErrorConstants {
 
     public static final String ERROR_CANNOT_EDIT_OTHER_USER_INFO = "error/cannot-edit-other-user-info";
+    public static final String ERROR_LOGIN_REQUIRED = "/error/login-required";
 
     private ErrorConstants() {}
 }
