+++ src/main/java/com/codessquad/qna/web/question/QuestionController.java
@@ -1,9 +1,9 @@
 package com.codessquad.qna.web.question;
 
-import com.codessquad.qna.common.constants.CommonConstants;
+import com.codessquad.qna.common.error.exception.LoginRequiredException;
 import com.codessquad.qna.common.error.exception.QuestionNotFoundException;
-import com.codessquad.qna.web.user.User;
 import com.codessquad.qna.common.utils.HttpSessionUtils;
+import com.codessquad.qna.web.user.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,11 +40,7 @@ public class QuestionController {
 
     @PostMapping("/questions")
     public String createQuestion(HttpSession session, @RequestParam String title, @RequestParam String contents) {
-        User loginUser = HttpSessionUtils.getUserFromSession(session).orElse(null);
-        if (loginUser == null) {
-            return CommonConstants.REDIRECT_LOGIN_PAGE;
-        }
-
+        User loginUser = HttpSessionUtils.getUserFromSession(session).orElseThrow(LoginRequiredException::new);
         Question question = new Question(loginUser, title, contents);
         questionRepository.save(question);
         return "redirect:/";
@@ -65,16 +61,12 @@ public class QuestionController {
 
     @GetMapping("/questions/{id}/form")
     public String goQuestionModifyForm(@PathVariable Long id, Model model, HttpSession session) {
-        User loginUser = HttpSessionUtils.getUserFromSession(session).orElse(null);
-        if (loginUser == null) {
-            return CommonConstants.REDIRECT_LOGIN_PAGE;
-        }
-
+        User loginUser = HttpSessionUtils.getUserFromSession(session).orElseThrow(LoginRequiredException::new);
         Question question = getQuestionIfExist(id);
+
         if (!question.isWrittenBy(loginUser)) {
             return "redirect:/questions/" + id;
         }
-
         model.addAttribute("question", question);
 
         return "questions/modify-form";
@@ -85,12 +77,9 @@ public class QuestionController {
                                  HttpSession session,
                                  @RequestParam String title,
                                  @RequestParam String contents) {
-        User loginUser = HttpSessionUtils.getUserFromSession(session).orElse(null);
-        if (loginUser == null) {
-            return CommonConstants.REDIRECT_LOGIN_PAGE;
-        }
-
+        User loginUser = HttpSessionUtils.getUserFromSession(session).orElseThrow(LoginRequiredException::new);
         Question question = getQuestionIfExist(id);
+
         if (!question.isWrittenBy(loginUser)) {
             return "redirect:/questions/" + id;
         }
@@ -103,14 +92,13 @@ public class QuestionController {
     @DeleteMapping("/questions/{id}")
     @Transactional
     public String deleteQuestion(@PathVariable Long id, HttpSession session) {
-        User loginUser = HttpSessionUtils.getUserFromSession(session).orElse(null);
-        if (loginUser == null) {
-            return CommonConstants.REDIRECT_LOGIN_PAGE;
-        }
+        User loginUser = HttpSessionUtils.getUserFromSession(session).orElseThrow(LoginRequiredException::new);
         Question question = getQuestionIfExist(id);
+
         if (!question.isWrittenBy(loginUser)) {
             return "redirect:/questions/" + id;
         }
+
         if (question.isDeletable()) {
             answerRepository.deleteAnswersInQuestion(question);
             questionRepository.save(question.delete());
