+++ src/main/resources/templates/error/login-required.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="kr">
+<head>
+  <meta charset="UTF-8">
+  <title>로그인해주세요.</title>
+</head>
+<body>
+<h1>해당 기능은 로그인해야 사용하실 수 있습니다.</h1>
+<script>
+  setTimeout(() => {
+    window.location.assign('/users/login');
+  }, 1000);
+</script>
+</body>
+</html>
